+++ packages/core/styles/src/packages/composite/split-button/index.ts
@@ -1,0 +1,5 @@
+import * as styles from "./nimbus-split-button.css";
+
+export const splitButton = {
+  ...styles,
+};