+++ packages/react/src/composite/SplitButton/src/components/SplitButtonAction/index.ts
@@ -1,0 +1,5 @@
+export { SplitButtonAction } from "./SplitButtonAction";
+export type {
+  SplitButtonActionProps,
+  SplitButtonActionProperties,
+} from "./SplitButtonAction.types";