+++ packages/react/src/composite/SplitButton/src/index.ts
@@ -1,0 +1,7 @@
+export { SplitButton } from "./SplitButton";
+export type {
+  SplitButtonProps,
+  SplitButtonProperties,
+  SplitButtonActionProps,
+  SplitButtonActionProperties,
+} from "./splitButton.types";